Test the openpi codecs from openpi's own tests
`test_remote_env.py` imported `vendors.openpi` to check that a LIBERO observation
encodes through the DROID codec — core reaching into a vendor. The check is about
those codecs, so it moves to `vendors/openpi/tests/`, taking its raw-obs fixture and
policy-input helper with it, and is named for the two codecs it actually exercises.

`test_keys.py`'s guard covers verdict keys as well as observation keys, so name it
for the set it checks. Trims the cross-target paragraph in the eval docs.
