+++ positronic/simulator/env_server/tests/test_remote_env.py
@@ -12,9 +12,8 @@ from positronic.cfg.eval.sim import positronic as native_cfg
 from positronic.cfg.eval.sim import robolab as robolab_cfg
 from positronic.dataset import Episode
 from positronic.dataset.local_dataset import LocalDataset
-from positronic.dataset.serializers import expand_suffixed
 from positronic.drivers.roboarm import command as roboarm_command
-from positronic.eval import Embodiment, Task
+from positronic.eval import Task
 from positronic.inference import main
 from positronic.policy import Policy, Session
 from positronic.policy.codec import ActionTimestamp
@@ -26,10 +25,8 @@ from positronic.simulator.env_server.proxy import RemoteEnvControlSystem
 from positronic.simulator.env_server.server import EnvProtocol
 from positronic.simulator.env_server.tests.conftest import serve_env
 from positronic.simulator.env_server.tests.mujoco_env import CAMERAS, make_mujoco_env, remote_stack_cubes_eval
-from positronic.simulator.libero.adapter import LiberoAdapter
 from positronic.simulator.robolab.adapter import RobolabAdapter
 from positronic.tests.testing_coutils import drive_scheduler
-from positronic.vendors.openpi import codecs as openpi_codecs
 
 
 class FakeRenderer:
@@ -304,40 +301,6 @@ def test_every_sim_eval_publishes_the_shared_camera_keys(eval_cfg):
     written for it. Every sim publishes the shared pair, whatever the benchmark calls those cameras."""
     observations = eval_cfg.instantiate().embodiment.observations
     assert {keys.EXTERIOR_IMAGE, keys.WRIST_IMAGE} <= set(observations)
-
-
-def _policy_inputs(embodiment: Embodiment, adapter: EnvAdapter, raw_obs: dict) -> dict:
-    """The observation a policy receives: the adapter's canonical signals through the embodiment's serializers,
-    unfolded into wire keys the way ``Harness._build_obs`` assembles them, plus the task the context carries."""
-    inputs = {keys.TASK: 'pick up the black bowl'}
-    for name, value in adapter.observations(raw_obs).items():
-        serializer = embodiment.observations[name].serializer
-        inputs.update(expand_suffixed(name, serializer(value) if serializer is not None else value))
-    return inputs
-
-
-# A LIBERO ``step`` payload, the shape ``LiberoEnv._observe`` returns over the wire.
-_LIBERO_RAW_OBS = {
-    'eef_pos': np.array([0.1, 0.2, 0.3]),
-    'eef_quat': np.array([0.0, 0.0, 0.0, 1.0]),
-    'joint_pos': np.zeros(7),
-    'joint_vel': np.zeros(7),
-    'grip': 0.4,
-    'agentview_image': np.zeros((256, 256, 3), dtype=np.uint8),
-    'eye_in_hand_image': np.zeros((256, 256, 3), dtype=np.uint8),
-    'sim_state': np.zeros(4),
-}
-
-
-@pytest.mark.parametrize('codec_cfg', [openpi_codecs.droid_obs, openpi_codecs.libero_obs], ids=['droid', 'libero'])
-def test_libero_observation_encodes_through_any_codec(codec_cfg):
-    """Serve once, score anywhere: a LIBERO observation encodes through a codec built for another env — the DROID
-    one — as readily as through LIBERO's own. What it scores is another matter; the viewpoints differ."""
-    ev = libero_cfg.spatial.instantiate()
-    inputs = _policy_inputs(ev.embodiment, LiberoAdapter(libero_cfg.spatial.kwargs['camera_dict']), _LIBERO_RAW_OBS)
-
-    codec = codec_cfg.instantiate()
-    assert set(codec.encode(inputs)) == set(codec.dummy_encoded())
 
 
 class _JointposChunks(Policy):
